improved performance of formatNumber fixes #4
FormatNumber is 4x quicker in terms of total time leading to an overall
increase of 7.69% to the program. It used to be 11% total execution time
and now is it 2.5%. This was achieved by implementing the [MDN](https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Number/toLocaleString) advice and replacing the most
expensive function in formatNumber.